Use service instances for multiple subreddits

+++ install.sh
@@ -1,19 +1,22 @@
 #!/usr/bin/bash
+export PIP=pip
 export SYSTEM=/etc/systemd/system
-#export SYSTEM=/tmp/test
 export BIN=/usr/local/bin
-#export BIN=/tmp/bin
 export LIB=$HOME/.local/lib/python-local/nasautils
-#export LIB=/tmp/lib
 export LOGDIR=/etc/logrotate.d
-#export LOGDIR=/tmp/logdir
 export SYSTEMCTL=systemctl
-#export SYSTEMCTL=echo
+# Uncomment below for debug
+# export PIP=echo
+# export SYSTEM=/tmp/test
+# export BIN=/tmp/bin
+# export LIB=/tmp/lib
+# export LOGDIR=/tmp/logdir
+# export SYSTEMCTL=echo
 
 # Next two save the local context so running as sudo can find them
 export home=$HOME
 export user=$USER
-pip install --upgrade -r requirements.txt
+$PIP install --upgrade -r requirements.txt
 
 ./install-db.sh
 
@@ -25,7 +28,7 @@ do
     export service
     sudo -E bash -c "sed 's@USER@$user@;s@HOME@$home@' $service.service  > $SYSTEM/$service.service"
     [[ -f $service.timer ]] && sudo cp $service.timer $SYSTEM
-    sudo cp src/$service.py $BIN
+    [[ -f $service.py ]] && sudo cp src/$service.py $BIN
 done
 
 sudo cp nasabot.logrotate $LOGDIR/nasabot
@@ -34,7 +37,15 @@ sudo $SYSTEMCTL daemon-reload
 for service in $(cat services)
 do
     [[ -f $service.timer ]] && service=$service.timer
-    sudo $SYSTEMCTL enable $service
-    sudo $SYSTEMCTL restart $service
+    if [[ $service == *"@"* ]] ; then
+        for subreddit in $(cat subreddits)
+        do
+            sudo $SYSTEMCTL enable $service.$subreddit
+            sudo $SYSTEMCTL restart $service.$subreddit
+        done
+    else 
+        sudo $SYSTEMCTL enable $service
+        sudo $SYSTEMCTL restart $service
+    fi
 done
